refactor(backends): Normalize outputter methods with self as first arg
This is useful for some outputters but not others, but it's hard to
implement something in one when none of the others (and hence all of the
calling code) is wrong. This is a pretty big change, but not too likely
to break most people's documents (it will break many of mine). I'll add
a check to error with an intelligent message rather than garbage if the
first argument isn't correct.

This means everything in the SILE.outputter space should be called as
a `:method()`. Period all the time.

This will eventually make the whole outputter thing easier to subclass
as well.