fix(packages): Align pullquote ending mark with outside margin
This will be a change in rendering but seems to have been the intention
all along, the way glues and reboxing were implemented was just bogus.
With this the opening and closing marks are both aligned to the
surrounding r/lmargin instead of the opening being outside and closing
inside.