ci(actions): Keep ‘latest’ and ‘vX’ tags updated
We previously had a Docker build system that retags the latest vX.Y.Z
release as ‘latest’. This moves that upstream to actual repo tags. This
way other systems can fetch and build ‘latest’ and get the latest stable
version. This will be a moving tag which is some trouble to fetch
locally, but experience with Fontship suggests this is the least
disruptive way to accomplish this.

Additionally new major version number tags are introduced to keep track
of major updates, so v0 is the latest release that isn't 1.x or newer.
This will make a *lot* more sense after we release 1.x and hence move
ourselves into true semantic versioning ranges.