refactor(classes)!: Convert class inheritance from stdlib to Penlight
BREAKING CHANGE: This changes the way classes are represented as Lua
objects and the mechanism used for inheritance. While shims will be in
place to catch most cases that use old syntax it is not possible to
grantee 100% API compatibility. If you have classes that do anything
remotely fancy (i.e. not just copy/paste from SILE examples) they may or
may not work at all; and even if they do they should be updated to
explicitly use the new API.