chore(tooling): Update license years and copyright holders
Without being a lawyer and posting sheets of legalise, there is always
going to be some ambiguity. Without a contributor agreement the question
is sometimes raised as to whether code snippets are copyright by their
respective authors and released under the project's license terms or if
the whole thing ends up being held by the stated copyright holder. For
a permissive license like MIT this is rarely an issue since it doesn't
materially affect most end user scenarios anyway. It's also possible to
argue that the copyright holder here is collective, "SILE Contributors".
In the absence of lawyerly advice, my tendency for open source projects
is to err on the side of explicitly stating multiple copyright holders
just to reduce the potential "bus factor" if something were to come up
and a question of who was authorized to speak on behalf of a project was
called into question.

Likewise copyright dating is usually assumed to be inclusive of actual
submission dates and is held in perpetuity anyway, but it doesn't hurt
to keep the date stamps current.