refactor(classes): Rewrite texlike command definitions in Lua
Having these in a Texlike typesetter at this particular location
introduces a race condition because the typesetter may not be
initialized. We could check for it, but we then limit how the class can
be used and in some use cases it is useful to initialize a class *then*
attach a typesetter to the class's default frame.