fix(docker): Work around libtexpdf build having side effects
Docker Hub and other remote builds work fine, but local builds can fail
if the host system is sufficiently different from the container being
built. The submodule "installs" files via a local copy as a side effect
of building, so detecting whether it has been built isn't enough to fix
the in-container config.