chore(packages): Handle Pandoc's space hacks (#1163)
This is a cheap hack to work around space hacks in non-Unicode document
formats. Long term getting the Pandoc Writer to output good Unicode
would be a better option, but there are some barriers to that happening
easily and working in all the places this hack is expected to work.