feat(measurements): Add ‘hm’ (himetric) unit
Metric equivalent of ‘twip’ in the SI world, adding largely for
posterity because it's low cost. I use mostly imperial units, but this
could be useful for the same reason twips are used instead of very small
decimals of points: working with smaller units and larger numbers is
easier mental gymnastics than very small fractions of large units.