grid.lua: Don't add a full gridSpacing where none would suffice
Previously when calculating the space toadd in makeUp, a cursor positioned exactly at a multiple of the gridSpacing would insert a full gridSpacing vglue, even when a zero vglue would also maintain proper spacing.  In other words, the range of possible values of toadd was (0, gridSpacing].

This change modifies the range of toadd to instead be [0, gridSpacing), which prevents the addition of unnecessary space.