fix(packages)!: Avoid unnecessary skips to next grid space
(Note: this is an empty commit to add a conventional-commits compatible
ChangeLog message to the release. The actual change is in d9abf6f1 and
thanks goes to @zavislak for the fix.)

BREAKING CHANGE: Previous to this release any and all leading between
paragraphs (as set with document.parskip) –even a 0 height skip– would
result in the skip of one full empty grid space — as if parskip had been
set to something approximating a full line height.  This change corrects
the calculation so if a 0 height skip is added and everything fits, the
next line or paragraph will continue uninterrupted in the next grid
slot.  To get the previous layout behavior back, document.parskip must
be explicitly set to be something larger than 0.  Even a minimal 1pt
skip will result in paragraph spacing that includes one full grid height
left blank as before:

    \set[parameter=document.parskip,value=1lh]