fix(packages): Don't over-isolate functions run in Pandoc imports
Many functions *want* to manipulate the settings environment for future
paragraphs. (Think a heading that sets \noindent for future paragraphs.)
Having everything isolated is nice it its own way, but that's not why
SILE is written in Lua.

This isolates the things that are clearly meant to be so but lets
everything else run (mostly) free.