fix(packages): Fix autodoc parsing, typeset string not series of bytes
This "magically" worked while stdlib was monkey patching Lua's string
handler because it allowed strings to be iterated like a table — as
a list of bytes. That's not what the `content` argument of SILE commands
is supposed to iterate on, but it was working until we stopped the
monkey business.