refactor(typesetter)!: Change typesetter instancing to Penlight model
BREAKING CHANGE: Making a new instance of the typesetter should now be
done by *calling* `SILE.defaultTypesetter()` instead of copying the
object. It has been changed from a std.object to a Penlight class. As
such the correct initialization function is also now `_init()` instead
of `init()`. A shim is in place to catch legacy usage, but this will be
removed in the future.