test(packages): Mark balanced frame test as known-bad
Recent package refactoring started flunking this test, so I looked into
it more. It turns out it was broken before and is actually *closer* to
right now, but still not all the way there. There was never enough
content on the first page to get anywhere close to the footnotes frame
with balanced columns enabled. Because the balancer was a noop before
(it was that broken) the second column was filling up. Now with the
frame break counting in the balancer neither column is big enough. Since
the frame balancer isn't fully working either it isn't clear to me how
to fix this test, so I'm at least marking it as bogus for now.