refactor(classes): Move :post_init() actually after init as :start()
The pseudo-post hack handled order — it could do things *later* inside init,
but it was never truly *post* init. This turns out to be limiting and causes
issues with legacy classes because access to the class from the SILE global is
not wired up yet. I toyed with other hacks, but it turns out if we just
*actually* run it after the class is in place we're all good. In the end
it looks like a good compliment to the existing :finish() method.