docs(manual): Update docs author to show more than Simon
I fiddled around a bit to try to come up with a fair way to represent
that there are several major docs contributors without making this
a reprint of `git shortlog -ns -- documentation`. This was the best
I came up with. The criteria I used was invitation to join the GitHub
org (we have one unaccepted open invite but since that fellow literally
wrote the documentation *package* I had to go with invite date rather
than accept date!), sorted by invite, and then checked if their
contributions included manual text or package documentation.