fix(packages): Fix coding errors in untested corners of bibtex package
Clearly nobody is using this except the demo on our website. Setting the
style was ostensibly a document command but it took a string argument
and documents can only pass table content, so ya. I didn't bother
marking the deprecation and refactoring as a setting as a breaking
change since it clearly wasn't working before.