test(packages): Avoid typesetter flip-flop
This test is supposed to be checking if color SVG fonts compose right in
TTB mode. Unfortunately the half-way-fix for being able to mix and match
color fonts and font fallbacks requires you use only one of those at
a time. The macro to switch fonts back and forth creates a mess of
shapers. Right now we can only use one or the other tool. At alternative
solution to just staying in one shaper as here is to
\font:clear-fallbacks to get rid of that shaper before using color
fonts.