fix(packages): Correct fall-back font processing
Now passes the correct font options along with the fallback so options
can be passed to fallbacks. Additionally outputs the glyph not found
error if it is actually needed with relative glyph. Corrects off-by-one
error causing shaping to run past the end of strings and throw a bogus
glyph-not-found error.

I had to refactor the entire stack. I think it's much more logical now
but there is probably one regression in the event a fallback font
provides shaping for a code point via more than one glyph.