feat(inputters): Promote Lua to first class input filetype, improve input type detection
This started out as a refactor to handle preamble content properly but
the current solution allowing execution of Lua code if parsing as SIL
was causing all sorts of weird results, so now it is just its own thing.

As a bonus this allows lots of things like Lua scripts that return ready
to process ASTs, functions that do whatever, strings that are their own
documents in any allowed format, etc.