refactor(core): Stop pretending 3rd party library is scoped to SILE
The SILE.fluent object is just an instantiation of fluent-lua with
nothing tying it to that scope or tracking document state or anything.
Hence the oddball usage of having to set_locale() every time we use it.
Perhaps it *should* be setup to track the document state, but this
refactor at least makes the status quo clear and makes room for a future
where the scope *is* related.