refactor(core): Move top level deprecation shims to their own home
It gets harder and harder to reason about what a file does and if it can
be moved or refactored when it is full of shims. This is a home for
shims to top level stuff not needed by SILE itself and just provided to
users at the last minute to ease their transition to new APIs.