refactor(core)!: Move registerCommand() out of global to classes
BREAKING CHANGE: The role of document commands has always been tightly
scoped to classes. For example the *book* class has a `\footnote`
command while *plain* does not—unless you manually load the package and setup the
frames. In spite of this obvious functional scope, registering commands
has been a global operation that stored them in a global registry. In
order to allow SILE to be used more programmatically as a library with
potentially more than one document and class being processed at at once,
these need to be moved out of the global scope. This will also
facilitate things like being able *unload* packages and revert to
previous functionality for anything they over-rood on load. For now the
functionality is shimmed, but code using the `SILE.registerCommand()`
function should switch to the method of the same name on the current
class, i.e. `class:registerCommand()`.