chore(i18n): Drop 'und' localizations
We *want* to fail with an error if localized strings are requested for
non-languages, so providing generic English stand-ins seems like
a counter-productive move.

Additionally for anybody starting a new localization it would be better
to start with a language similar to their target because implementation
details might differ. With simple key/value lookups it might seem like
any language could be a template for any other, but Fluent is much more
flexible than that.