fix(math): Fix debug logs in TeX-like parsing
Make the debug statements in `texlike.lua` actually print trees instead
of just “table: <address>”. In the MathML debug log, the tree is
serialized back to a SILE command.

Change `compileToMathML` so that, in debug mode, the tree does not get
printed at each recursion step, but only at the last step.