refactor(packages): Drop class arg on package constructor, normalize with other modules
This is a small step backwards in scope, but it normalizes how all
different module types are instantiated so packages are not a special
case with an extra argument. Realistically we didn't have the class
scope isolated anyway since it was still used in the global, and long
term there are better ways to pass _ENV to functions that we should be
able to use to fix that scope issue. Even if we don't we would want to
change *all* module types, not just one.