feat(math): Express lengths in “mu” (math units)
As in TeX, 1 mu is defined to be 1/18th of the math font em unit, while
1 em stays relative to the main font size. Note that enspace, quads and
qquads are defined in terms of ems, so they stay relative to the main
font size even inside formulas (this causes a change in the
`math-spaces` regression test).

(See the TeXbook, pages 166 to 168.)

Standard spaces definitions (thin, med, thick) are moved to the
`elements.space` class (replacing the `newStandardHspace` function) to
be reused in the macro definitions of `texlike.lua`.

This also fixes spacing size in script and scriptscript modes, as
previously the size was scaled down twice, once in the call to
`newStandardHspace` and once in `elements.space:shape()` (this causes a
change in the `math-bigops` regression test).