fix(math): Fix caching of getMathMetrics
getMathMetrics’ cache is now per-font, using the key function defined in
SILE.fonts. This fixes the rendering of formulas in documents that have
different math font sizes. The effect is particularly noticeable in the
position of subscripts and superscripts.

getMathMetrics is moved to a member function of the mbox class to
avoid duplicating the logic that computes the `options` table. The
`options` table is renamed to `font` to avoid a clash with a different
table that has the same name in `elements.table`.