fix(packages): Parse height argument to `\raise` / `\lower` as measurement
This is an old function that predates our SILE.length / SILE.measurement
logic and units like `%fh` that are relative to the frame or page. It
was being parsed as a frame dimension, which I assume was just a way to
get user friendly string input and perhaps some expressiveness.
I couldn't actually think of a use case for making this a full frame
string and not accepting our standardized measurement types comes with
lots of downfalls especially in programmatic usage.

I'm not marking this as breaking because in most cases I think the new
behavior will be expected and I doubt the edge cases where they differ
are seeing much use.