chore(deps): Make pdf_metadata return int to silence GCC warning
The function is non-void, but it doesn't return anything in the end.
Without this, gcc complains with a "return reaches end of non-void
function" warning, which is treated as an error on certain Linux distros
such as openSUSE.

As noted by Omikhleia, this is the right fix as it returns number of Lua
values returned to the Lua calling context.