fix(classes): Place gutter margin to *left* side of odd pages
The default book class has had a layout that suggests a gutter margin,
but the margin has been on the wrong side of the isle since its
conception. Since the default language and layout is LTR and the default
first page is odd and the book class uses the default frameset as the
right master, it seems to me the wider margin should be on the left
where a LTR book binding would traditionally be.