fix(inputters): Relax SIL format sniffing to allow valid syntax
This allows the content sniffer to succeed and tag content as SIL when
the syntax is valid even if the document is invalid (e.g. not wrapped in
a single top level document tag). This helps with correctly identifying
STDIN streams that happen to be SIL and moves the ball down the road.
Our later parsing step basically redoes the same work but then has
better checks for a valid AST tree and hence better error messages in
the event of a malformed document.