fix(classes): Apply page/framebreak in hmode but warn the user
The implementation was somewhat broken (pagebreak did nothing, and
framebreak was a line break, if they were called in horizontal mode.
They now terminate the paragraph, and do their expected vertical
behavior... It probably better meet user expectations, but there are
dragons here, so we also issue a warning.

Various comments added to the code for consideration if need be at some
points.