test(packages): Update test for counters
Update deprecated calls.

Remove display formats that were undocumented and mixing the
concepts of numbering system and format style. Slighty breaking
possibly, but if this was supposed to be a real feature, a syntax
extension to counters would be better...

Note that example in arabic-indic was using decimal ١٬٩٨٤ and
is now a default number ١٩٨٤. Such a huge value is unlikely to be as
common case for counters... But this seems more consistent e.g. with
what CSS "list-style: arabic-indic;" also does, so likely better.