refactor(shapers): Create Harfbuzz font once
Creating HarfBuzz fonts was all over the place, done in a slightly
different way each time and with potential memory leaks (e.g. often
destroying the font, but not the face or the blob).

It was also reading the whole font data in Lua, then passing it to
HarfBuzz, which is inefficient—especially for large fonts. The code is
now using hb_blob_create_from_file() instead which will mmap the file
whenever it can.

The created hb_font_t is now cached in the Lua font and retrieved from
there when needed.