chore(deps): Require HarfBuzz >= 2.7.4
We were not requiring any specific version, but it is unlikely that the
code works with such old versions of HarfBuzz. Instead, require the
newest version of HarfBuzz we can get away with, the version available
in stable Debian/Ubuntu.

Remove checks for older versions of HarfBuzz and consistently use the
same font functions everywhere.

Update CMake build to also use HarfBuzz 2.7.4 and build it without
FreeType support as it is no longer needed. Also drop ragel dependency,
since ragel generated sources are included in the repository.