feat(typesetters): Implement hbox building logic in the typesetter
Construction of an hbox from content depends a lot on the
typesetter internal logic and structure. The implementation
here is based on the original one from the plain class, but
doesn't push the hbox into the typesetter queue (so that it
doesn't have to be "stolen" back if only built for measuring
it) and additionally extracts and returns migrating nodes
that would be lost otherwise inside the hbox.