docs(manual): Add a "maturity" indicator to packages in chapter c05
Re-organizing the topics in dedicated sections, we lost the big
dichotomy from earlier versions of the manual, between main
packages (that were sorted alphabetically) and experimental
packages (a unsorted and actually non-curated list of things that
range from usable to broken...).
We can do better, with a nicer 3-level indicator: it's neat, not
too intrusive, and respectful for the readers as to what they may
expect.