chore(deps): Document current MSRV in Cargo.toml
Documenting it should help in the future analizing what release tags
should be and whether or not to dump dependencies.

Technically the current limit would be 1.64.0, a limitation inherited
from `clap_builder` and `clap`. While we could adapt to support that
version, the linting rules we are using require 1.71.0. Since we're just
coming out with Rust support I don't feel bad at all about making it
a pretty new version. Hopefully that will give us a little breathing
room before we have to do major bumps to support MSRV bumps.