chore(classes)!: Remove obsolete/broken native markdown class
Drop the markdown parser based on a Lunamark fork. 3rd party packages
now support Markdown parsing in a more robust way than the vendered fork
of a Lua parser ever did.

Previously implemented in 5f462836f, but not marked as breaking. This
commit is to get the change to show up in breaking changes.