feat(packages)!: Use OpenType +ruby feature for ruby, toggle with setting
BREAKING CHANGE: The default rendering of Ruby readings has changed from
just using a bold weight to using the OpenType +ruby feature. Fonts that
support this should work with no change, but documents rendered in fonts
that do not support it will neet to set the `ruby.opentype` feature to
`false` to get the same rendering method as before.