Landscape option added to base class
An orientation option was declared in the base class.
When it is "landscape", the values of x and y in papersize are swapped.

Just to explain what maybe could look redundant in classes/base.lua:92:

While testing this feature, it worked well for several times, but sometimes
- without touching the code - it suddenly stopped to work!

The problem is that the option.orientation must be passed as parameter
in SILE.papersize and in those times when it didn't work, its value was nil
even though it actually worked many times before without any
modification.

After some time I found out why: since there is no order in a loop using
pairs(), sometimes the loop in class:setOptions() was getting orientation
as the last element, so when SILE.papersize was called, the value of
options.orientation was indeed nil, so the solution should be just
set the orientation value to class.options before papersize.

modified:   classes/base.lua
modified:   core/papersize.lua
new file:   tests/landscape.sil