feat(packages)!: Make most package _init() actions run only once
BREAKING CHANGE: The current (pseudo) idempotent behaviour when loading
a package potentially clobbers anything that has been modified
since the last load. Loading a package, then modifiying a function it
provides, then loading the same package again will clobber the
modifiecation. This is good for idempotency but not very good for user
experience when you may not be modifiying all aspects of a document
render pipeline at once, as in when using templates.

This change makes the default behaviour to run setting, raw handler, and
command registrations only once. An altertanive to :loadpackage() called
:reloadpackage() can be used to force all these registrations to be
rerun when the goal is to make sure of a specific state.