chore(core)!: Change default paragraph indent to 1 baseline skip
BREAKING CHANGE: The previous default paragraph indent was hard coded
with a point size (20pt) that did not adapt will to different fonts or
page layouts. The new default uses a relative unit that will adjust
based on the leading (1bs). This will cause most documents to reflow. To
keep them the same the setting may be reset to the old default either in
the document:

    \set[parameter=document.parindent,value=20pt,makedefault=true]

...or even from the CLI when rendering a document:

    $ sile -e 'SILE.settings:set("document.parindent", "20pt", true)'