chore(core)!: Change default space width to 1 space
BREAKING CHANGE: For ... reasons ... the default width of spaces in SILE
has been a hightly opinionated and non-standard 1.2 spaces. While it can
be argued that this makes some fonts and some documents look better, it
is a very strange thing to have as a global default. Unfortunately
setting it back to a more conventional 1 space is a major change and
will cause many/most documents to reflow.

The old default can be recovered either in documents with:

    \set[parameter=shaper.spaceenlargementfactor,value=1.2,makedefault=true]

...or even from the CLI when rendering a document:

    $ sile -e 'SILE.settings:set("shaper.spaceenlargementfactor", 1.2, true)'