chore(tooling): Drop obsolete Homebrew formula generator
This is cool but not very useful because it is constantly out of sync
with how upstream mantains the file. Style changes, dependency changes,
Ruby function changes, etc. all affect this no to mention our own local
build changes. These are harder to backport to the perl than it is just
to manually update the upstream formula.