chore(build): Replace AX_ prefix hijack with custom
These macros are identical across a bunch of projects I work on, hence
the naming scheme. I eventually decided it was confusing to know which
macros came from the autoconf-archive and which were my own, so not
using the AX_ prefix except for things that could be dropped and used
directly from the archive just makes sense.