fix(classes)!: Reset current parindent at end of paragraphs, not beginning
BREAKING CHANGE: Hitherto SILE has cleared the current.parindent setting
as soon as it used it at the beginning of a paragraph. With this
release, the setting is not being cleared until a paragraph is
explicitly ended. This will not have an affect on many documents, but
could completely blow up layout code that implicitly relied on the
effect. Normal paragraphs (e.g. separated by a blank line in the input)
and any use cases that explicitly called far ending a paragraph (e.g. by
calling `\par`) will be unaffected. But anywhere a paragraph break was
simulated by adding vertical space, the indentation will not be applied
the next start of a line. This means that calling any variant of `\skip`
inline in a paragraph will result in content beginning on a new line
without using the parindent setting.

Fixing this change in behaviour requires either explicitly resetting the
current.parindent setting *after* it is initially used or explicitly
ending a paragraph before or after placing a vertical skip.