fix(packages): Change retrograde package so 'target' version works as expected
The retrogade package as released in v0.15.0 was counter intuitive in
that the 'target' argument was the first version that had settings one
wanted to reset. It is more intuative to use if the target is the
version the document settings were designed for. In other words to undo
default changes released in v0.15.0 a document can specify a target of
v0.14.17. This would be a breaking change if the original hadn't been
released this morning and the original counterintuative behaviour had
any change of being deployed anywhere.