feat(packages): Support `@string` syntax in bibTeX bibliography
Support `@string` blocks in a bibtex file, and the use of the
thus defined strings in other entry fields.
Supports concatenation with the "#" operator.

As part of the syntax clean-up, all content between entries
is now skipped (it doesn't need to start with a percent sign,
everything that is not an entry is a comment.

See #2051