fix(build): Do not distribute main man page, requires Rust tooling
Requiring us to download Rust dependencies and build the whole Rust CLI
just to generate the source dist tarball is a waste. Distributing the
man page in the distribution package makes sense when generating it
requires tooling the user may not have, but in this case it *will* get
rebuild when they generate the CLI anyway.