feat(outputters): Add AST backend that dumps the abstract syntax tree
This also comes with the ability for *any* outputter to pre-process the
AST tree if it wanted to (e.g. for remove leaf types it would not be
able to output. If the outputter doesn't do anything to the AST it is
just a pass-through step.

Note this does *not* yet walk the AST and do any AST munging that
functions might engage in, it only catches the top level tree as parsed
by the inputter.

Additionally it is not a full serialization of the AST that could be
read back in with an input driver. We'd need a new dependency on
something like serpent for that, which would be a lovely next step.