feat(core): Add px unit as 0.75pt as per CSS3
We long did without it, but MathML examples may contain lengths in px.
And we'll be happy too with other uses cases such as Markdown etc.
This might break some code as SILE.types.length can ignore the unknown
unit and the raw number value would thus be regarded as being in pt.
But that's rather a parsing bug anyway, and SILE.types.measurement spits
an error, so let's not care.