fix(math): Stretchy symbols should work on large content (workaround)
The current stretchy "reshaping" picks the closest symbols in the math
glyph variants, but when none exist or they are still not close enough
(esp. far too small), OpenTypeand MathML Core say we should rely on
complex glyph assembly rules from glyph parts.
We don't have that yet: Here we fallback on re-scaling the glyph.
It might not look always good, but at least it's stretched as it
should around the content. It's a dirty compromise until something
better can be done.