test(math): Adjust MathML tests in math-ops.sil to the operator dictionary
The MathML code was using U29F5 (reverse solidus), and the supposedly
identical TeX-like code uses "setminus", which formerly mapped it to
the reverse solidus. It was changed in unicode-math in 2023 to correctly
map to U2216 - and our new operator dictionary table follows that
change. So the test is adapted to use U2216 too now, and the test
results also need to be regenerated.
On that occasion, the MathML code is also ajusted to use `<mrow>`
elements in sub-expressions, so the sizes of parentheses is in line
with the TeX-line examples.