refactor(core): Move backend selector from top level to to inputs table
The desired backend when using SILE as a library should now be set via
SILE.input.backend instead of SILE.backend. This makes the namespace
more consistend with other input parameters vs. the resulting loaded
modules.

This is technically a breaking change but will not be released in
a major version because the compatibility shim is expected to cover this
for the whole next major release cycle, so the breakage will only come
in v0.17.0.