refactor(build): Adapt custom macro for upstreaming to autoconf-archive
Since I seem to use this on every project I touch now, I went ahead and
spruced it up in a way that should make it acceptable as a contribution
to autoconf-archive. As of tihs writing it has been PR'ed but not
morged, so consider the exact script with this serial still fungable.

https://github.com/autoconf-archive/autoconf-archive/pull/324