chore(packages): Suppression of empty CSL macros is not that simple
Previous commit "Suppress empty CSL macros" did it right for several
styles (ABNT, etc.) but is a bit too brutal with other styles
(e.g. acta-philosophica) as a suppressed macro (that is, referring
to variables that are all empty) should also be tracked as an
empty variable.

Marked as "chore" as this didn't go in any release yet and to not
affect styles which were already working at this date.