feat(packages): Support the position condition in CSL engine
Quite useful for "note" citation styles (for "cit.", "ibid." and the
likes) although some "in-text" citation styles also benefit from it.

Here we just add the support to the CSL engine, but as with the
locator and the citation number, the position has to be provided by
the upper processor, depending on the context (i.e. previously cited
entries).