chore(core): Mark internal status flag as private
Technically this *should* be considered breaking, but I'm not
documenting it as such because it was only ever used (and intended for
use) internally, and only for developer tooling at that. This status
flag enable us to track whether we should consider the results of an
integration test a pass or fail even if the output varies due to
unsupported library conditions. It should be irrelevant to end users.