chore(core)!: Require all module methods to be called with a reference to self
BREAKING CHANGE: All functions belonging to all top level modules
(classes, packages, typesetters, shapers, inputters, outputters, etc.)
are to be called with a reference to self as the first argument
(typically with `module:method()` convenience syntax). Previously we
allowed some methods to circumvent this because they didn't need access
to anything else in the parent module. However they did frequently reach
into globals such as `SILE` which will not be supported in the future.
The only way they will have to reach out to the rest of the engine will
be through their parent module. This includes access to the typesetter,
settings, etc.