feat(packages)!: Stop supporting the legacy bibliography implementation
BREAKING CHANGE: The new CSL-based bibliography engine was introduced in
SILE 0.15.7 at the end of Nov. 2024, foreshadowing the eventual deprecation
of the earlier home-grown legacy implemention of (a subset) of the Chicago
author-date style.
It's now removed, so we can move forward simplifying the code base and
extend the package with new features.
The bibtex.style setting is however kept, so as not to break documents
or modules that were setting it (to 'csl' or 'chicago'), but it does
nothing now, and the CSL implementation is always used.