feat(packages): Split CSL processing logic out of the bibtex package
It started as a refactor, but ends up as a feature:
Bibliography processing is now available in the CslProcessor class.
The latter may be used outside of SILE's regular processing flow
(that is, using SILE as a Lua-on-steroids toolkit with ICU, etc. but
without going through the typesetter etc. - for instance, I am using
it to generate HTML versions of my bibliographies. It's still rough,
but paves the way for better abstractions).
The package itself now just implements commands for SILE as a
typesetting engine, delegating to that processor and just containing
the very specific code layer for processing in documents.