refactor(packages): Wrap bibliography entry output in XML paragraph blocks
Introducing <bibParagraph> wrapping entries, instead of a <par>
in-between. The way it's implemented does not change anything as far as
SILE processing is concerned, but makes it easier for other output
processors (e.g. to HTML where one would want explicit start-end
constructs to mark paragraphs and translate them to anything suitable
(<p> or <div>) with appropriate CSS.