fix(packages): Rework how sorting and substitutions are done in bibliographies
In the commits in PR 2230, that went in SILE 0.15.13, I tried to fix an
issue regarding the suppression of empty macros (impacting substitutes
in names). It turned out to cause other issues (2283) and badly break
some previously working styles (Chicago, MLA) with some entries...

This is an attempt at repairing them. But something still eludes me in
the CSL specifications, and while it does seem to work now with the
styles we previously tested, using the new Chicago styles recently
updated in the CSL repository shows that something is still amiss. These
updated styles have a lot more nested macro with conditionals than
previously, and I haven't pinpointed where the misinterpretation or raw
bug could lie. Yet, the code as now refactored makes more sense, with
less side effects and better scoping, so at least it's worth having,
albeit perhaps as a partial fix...