test(packages): Remove obsolete legacy bibtex tests
Since the removal of the legacy bibliography implementation, these two
tests were deprecated.
One (bug-2054.bib) is clearly obsolote (checking a side-effect of fluent
in bibligraphy internationalization, but the new implementation uses
CSL)
The other leads to a wholly different output (really following Chicago
Author-Date now) but would have been hard to adapt, and was also fairly
uncomplete. A different testing approach would be required, as we move
to a more layered and modular approach.