fix(types,fonts): Base "ex" unit on proper metrics and add "cap" unit
The "cap" unit is proposed in CSS4 Draft and makes sense for cap-height
font adjustement.
Both "ex" and "cap" shall be based on font metrics when available, and
only measure a character as a fallback heuristic.
Some fonts may not have the necessary Latin characters, in which case
we shall resorti to a decent fallback (as per CSS recommendations).