feat(packages): Add scaling and anchoring to the background command
New options: page (for multi-page images such as a PDF), scale (for
disabling scaling), preserveaspect (for aspect ratio preservation when
scaling), anchor (for positioning of the image when needed).

Moreover a color and an image can both be set, for cases where the image
has transparent areas or isn't scaled to the full bleed area.